split structs across headers
