+++ fwdpy11/headers/fwdpy11/discrete_demography/forward_demes_graph/asymmetric_migration.hpp
@@ -1,0 +1,16 @@
+#pragma once
+
+#include <cstdint>
+#include "demes_model_time.hpp"
+
+namespace fwdpy11
+{
+    namespace discrete_demography
+    {
+        struct AsymmetricMigration
+        {
+            std::int32_t source, dest;
+            demes_model_time start_time, end_time;
+        };
+    }
+}
